Refactor NeuriteElement class to support basal and apical dendrites
This commit modifies the NeuriteElement class to support the identification and manipulation of basal and apical dendrites. It adds new methods `IsApicalDendrite()` and `IsBasalDendrite()` to check the type of neurite element, and `SetApicalDendrite()` and `SetBasalDendrite()` to set the type. Additionally, the `GetIdentifierSWC()` method is updated to return the appropriate structure identifier for basal and apical dendrites.